System-Design/phb-lab/worker/requirements.txt
minio==7.1.16
pika==1.3.2
ffmpeg-python==0.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
